Fix eval cap bug, minimax overflow, date provenance; add tests
Correctness and quality fixes from the code review:

- #9: eval mode no longer caps the tree walk at worst_case_depth, which
  misreported valid-but-deep paths as FAIL. Introduce a shared walk_target()
  helper (database.hpp/cpp) used by both build_db::evaluate and CLI eval, with
  a generous WALK_DEPTH_CAP and explicit MissingEdge/ExceededCap statuses.
- #23: deduplicate the tree-walk loop between build_db and main via walk_target.
- #14: minimax_inner count array widened uint8_t -> uint16_t to remove latent
  overflow if called with a large candidate set.
- #20: derive words_date from the words-file mtime (POSIX stat) with a --date
  override, instead of hardcoding the date string.
- #25: correct CLAUDE.md note — the repo is public by design.

Tests (43 -> 48):
- Real post-finalize tampering test for verify_integrity (#21): mutate a node
  row via raw sqlite3, confirm the stale checksum is detected.
- walk_target depth/missing-edge coverage.
- minimax_best_guess unit tests (single/empty sets, optimal-depth cross-check
  against greedy, achievable-bound check).

+++ src/solver.cpp
@@ -233,7 +233,11 @@ EntropySolver::minimax_inner(std::span<const uint16_t> candidates,
         if (prune_at <= 2) return DEPTH_IMPOSSIBLE;
 
         // First pass: count how many candidates fall in each pattern bucket.
-        std::array<uint8_t, PATTERN_COUNT> counts{};
+        // uint16_t (not uint8_t): although the only production caller gates
+        // minimax to candidate sets ≤ MINIMAX_THRESHOLD, this is public API and
+        // the top-level counts array sees the full candidate set. uint16_t
+        // cannot overflow for any list WordList::load accepts (≤ 65,535 words).
+        std::array<uint16_t, PATTERN_COUNT> counts{};
         for (uint16_t ai : candidates) {
             ++counts[patterns_.get(gi, ai)];
         }
